+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 **July 14, 2023 - [DESCW-871](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-871)**
 - Frontend
     - Updated admin contact section with new table component
+    - Updated admin supplier section with new table component
 
 **July 13, 2023 - [DESCW-1354](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1354)**
 - Backend